distributed to folders
