+++ ContactsApp/operations/getContact.js
@@ -1,0 +1,9 @@
+const readContent = require("./readContent");
+
+const getContact = async (contactId) => {
+  const contacts = await readContent();
+  const [contact] = contacts.filter((contact) => contact.id === contactId);
+  return contact;
+};
+
+module.exports = getContact;
