+++ ContactsApp/operations/index.js
@@ -1,0 +1,10 @@
+const listContacts = require("./listContacts");
+const getContact = require("./getContact");
+const addContact = require("./addContact");
+const removeContact = require("./removeContact");
+module.exports = {
+  listContacts,
+  getContact,
+  addContact,
+  removeContact,
+};
