+++ ContactsApp/operations/listContacts.js
@@ -1,0 +1,8 @@
+const readContent = require("./readContent");
+
+const listContacts = async () => {
+  const result = await readContent();
+  return result;
+};
+
+module.exports = listContacts;
